refactor: split problem-service admin.ts (3318→35 lines)
Extract 9 focused route modules from the monolithic admin.ts:
- admin-events.ts (498 lines) — Event CRUD
- admin-problems.ts (500 lines) — Problem CRUD + marketplace
- admin-import-export.ts (328 lines) — Import/export
- admin-contests.ts (227 lines) — Contest management
- admin-teams.ts (81 lines) — Team management
- admin-dashboard.ts (112 lines) — Dashboard/stats
- admin-templates.ts (502 lines) — Template CRUD
- admin-ai.ts (364 lines) — AI generation
- admin-deploy.ts (731 lines) — AWS + GameDay deploy
- admin-shared.ts (75 lines) — Shared repos/auth

admin.ts is now a 35-line aggregator mounting sub-routers.
No route paths, schemas, or logic changed.
